Rename a class, remove unnecessary roles.
