--- config/roles/api/admin.php
@@ -1,145 +1,0 @@
-<?php
-
-return [
-    'custom' => [
-        'auth_profile.edit',
-        'record_menu.access',
-        'engage_menu.access',
-        'case_menu.access',
-        'support_menu.access',
-        'settings_menu.access',
-        'kb_menu.access',
-        'support_item.create',
-        'support_item.edit',
-        'support_item.show',
-        'support_item.delete',
-        'support_item.access',
-        'support_training_item.create',
-        'support_training_item.edit',
-        'support_training_item.show',
-        'support_training_item.delete',
-        'support_training_item.access',
-        'support_feedback_item.create',
-        'support_feedback_item.edit',
-        'support_feedback_item.show',
-        'support_feedback_item.delete',
-        'support_feedback_item.access',
-        'support_page.create',
-        'support_page.edit',
-        'support_page.show',
-        'support_page.delete',
-        'support_page.access',
-        'record_enrollment_item.show',
-        'record_enrollment_item.access',
-        'record_program_item.show',
-        'record_program_item.access',
-        'record_student_item.show',
-        'record_student_item.access',
-        'prospect_menu.access',
-        'report_menu.access',
-        'report_student.create',
-        'report_student.edit',
-        'report_student.show',
-        'report_student.delete',
-        'report_student.access',
-        'report_prospect.create',
-        'report_prospect.edit',
-        'report_prospect.show',
-        'report_prospect.delete',
-        'report_prospect.access',
-        'journey_menu.access',
-    ],
-
-    'model' => [
-        'auditLog' => [
-            '*',
-        ],
-        'caseItem' => [
-            '*',
-        ],
-        'caseItemPriority' => [
-            '*',
-        ],
-        'caseItemStatus' => [
-            '*',
-        ],
-        'caseItemType' => [
-            '*',
-        ],
-        'caseUpdateItem' => [
-            '*',
-        ],
-        'engagementEmailItem' => [
-            '*',
-        ],
-        'engagementInteractionDriver' => [
-            '*',
-        ],
-        'engagementInteractionItem' => [
-            '*',
-        ],
-        'engagementInteractionOutcome' => [
-            '*',
-        ],
-        'engagementInteractionRelation' => [
-            '*',
-        ],
-        'engagementInteractionType' => [
-            '*',
-        ],
-        'engagementStudentFile' => [
-            '*',
-        ],
-        'engagementTextItem' => [
-            '*',
-        ],
-        'institution' => [
-            '*',
-        ],
-        'journeyEmailItem' => [
-            '*',
-        ],
-        'journeyItem' => [
-            '*',
-        ],
-        'journeyTargetList' => [
-            '*',
-        ],
-        'journeyTextItem' => [
-            '*',
-        ],
-        'kbItem' => [
-            '*',
-        ],
-        'kbItemCategory' => [
-            '*',
-        ],
-        'kbItemQuality' => [
-            '*',
-        ],
-        'kbItemStatus' => [
-            '*',
-        ],
-        'prospectItem' => [
-            '*',
-        ],
-        'prospectSource' => [
-            '*',
-        ],
-        'prospectStatus' => [
-            '*',
-        ],
-        'recordEnrollmentItem' => [
-            '*',
-        ],
-        'recordProgramItem' => [
-            '*',
-        ],
-        'recordStudentItem' => [
-            '*',
-        ],
-        'supportPage' => [
-            '*',
-        ],
-    ],
-];
