--- config/roles/web/user.php
@@ -1,6 +1,0 @@
-<?php
-
-return [
-    'custom' => [],
-    'model' => [],
-];
